removed warnning message for build

+++ src/drivers/DISCOVERY_STM32F7/src/vl53l0x_platform.c
@@ -51,7 +51,7 @@ int _I2CWrite(VL53L0X_DEV Dev, uint8_t *pdata, uint32_t count)
    status                                                = HAL_I2C_Master_Transmit(Dev->I2cHandle, Dev->I2cDevAddr, pdata, count, i2c_time_out);
    if (status)
    {
-      debug_output_error("I2C error 0x%x %d len \r\n", Dev->I2cDevAddr, count);
+      debug_output_error("I2C error 0x%x %d len \r\n", Dev->I2cDevAddr, (int) count);
    }
    return status;
 }
@@ -69,7 +69,7 @@ int _I2CRead(VL53L0X_DEV Dev, uint8_t *pdata, uint32_t count)
    status                                                = HAL_I2C_Master_Receive(Dev->I2cHandle, Dev->I2cDevAddr | 1, pdata, count, i2c_time_out);
    if (status)
    {
-      debug_output_error("I2C error 0x%x %d len \r\n", Dev->I2cDevAddr | 1, count);
+      debug_output_error("I2C error 0x%x %d len \r\n", Dev->I2cDevAddr | 1, (int) count);
    }
    return status;
 }
